COMMIT - Concluindo a parte de validação dos métodos que o usuário tem acesso.

+++ MP.ApiDotNet6.Domain/Entities/UserPermission.cs
@@ -4,18 +4,18 @@ namespace MP.ApiDotNet6.Domain.Entities
 {
     public sealed class UserPermission
     {
-        public UserPermission(int userId, int permissionId)
+        public UserPermission(long userId, long permissionId)
         {
             Validation(userId, permissionId);
         }
 
-        public int Id { get; set; }
-        public int UserId { get; set; }
-        public int PermissionId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
+        public long PermissionId { get; set; }
         public User User { get; set; }
         public Permission Permission { get; set; }
 
-        private void Validation(int userId, int permissionId)
+        private void Validation(long userId, long permissionId)
         {
             DomainValidationException.When(userId == 0, "Id do Usuário deve ser informada");
             DomainValidationException.When(permissionId == 0, "Id da Permissão deve ser informada");
